Ensure that ordered indexes get the appropriate score.
It wasn't working for a date attribute for me, because the @order wasn't set. Instead, it tried to extract the score from the stringified date, because Index#read_normalized returns the attribute as a string.  Thus, a date like "2011-12-03" would get scored as 2011 because the code would #to_i it.  By setting @order, it ensures that Ordering#score re-reads the value and gets a date object.