Make IndexSet lazy-retrieve the Redis client
Since we now do a call to the server when the client is created, it's
important to not create the client until we're actually ready to use it.

This is critical since Relix gets called during class initialization,
and backing resources may not have been set up yet. An example is a test
Redis server that needs to be booted but isn't set up until after
classes have been loaded.