Properly reset current values hash
If the condition upon which an index is based changes from true to false
and then back to true, we now handle that case, even when the model has
more than one index. We now always remove the old current values before
doing the mass set of the new one.

Closes #14.